pyproject.toml
[project]
name = "metasketch"
version = "0.1.0"
description = "AI Web Template with FastAPI, Jinja2, HTMX, TailwindCSS, and async SQLAlchemy."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "jinja2>=3.1.3",
    "pydantic>=2.7.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy[asyncio]>=2.0.30",
    "uvicorn[standard]>=0.29.0",
    "aiofiles>=23.2.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose>=3.3.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.19.0",
    "python-multipart",
    "python-jose>=3.3.0",
    "markdown>=3.5.2",
    "markupsafe>=2.1.3",
]

[tool.ruff]
line-length = 88

[tool.black]
line-length = 88

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = { find = { where = ["."] , exclude = ["documentacion"] } }
